requirements.txt
nltk
pandas
torch
pytorch_pretrained_bert
scikit-learn
numpy
sentence-transformers
datasets
transformers
en-core-web-trf
torchvision
layoutparser
spacy
git+https://github.com/facebookresearch/detectron2.git@v0.5#egg=detectron2
layoutparser[layoutmodels]
layoutparser[ocr]
layoutparser[paddledetection]
layoutparser[effdet]
spacy[transformers]
pillow
pdfplumber
pdf2image
tqdm
pybtex
python-docx
